+++ src/components/Projects/Terminologies/Terminology.scss
@@ -108,6 +108,7 @@
 
 .modal_search_result {
   display: flex;
+  flex-direction: column;
   align-items: center;
   justify-content: space-between;
   margin: 0 10px 10px 6px;
